+++ CODEOWNERS
@@ -62,7 +62,7 @@
 /bundles/org.openhab.binding.helios/ @kgoderis
 /bundles/org.openhab.binding.heliosventilation/ @ramack
 /bundles/org.openhab.binding.heos/ @Wire82
-/bundles/org.openhab.binding.heliosventiremote/ @ramack
+/bundles/org.openhab.binding.heliosventilation/ @ramack
 /bundles/org.openhab.binding.homematic/ @FStolte @gerrieg @mdicke2s
 /bundles/org.openhab.binding.hpprinter/ @cossey
 /bundles/org.openhab.binding.hue/ @cweitkamp
