+++ bundles/org.openhab.binding.heliosventilation/src/main/java/org/openhab/binding/heliosventilation/internal/HeliosVentilationBindingConstants.java
@@ -12,16 +12,8 @@
  */
 package org.openhab.binding.heliosventilation.internal;
 
-import java.io.IOException;
-import java.net.URL;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Properties;
-
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * The {@link HeliosVentilationBindingConstants} class defines common constants, which are
@@ -34,56 +26,18 @@ public class HeliosVentilationBindingConstants {
 
     public static final String BINDING_ID = "heliosventilation";
 
-    public static final String DATAPOINT_FILE = "datapoints.properties";
-
     // List of all Thing Type UIDs
-    public static final ThingTypeUID THING_TYPE_HELIOS_VENTILATION = new ThingTypeUID(BINDING_ID, "ventilation");
+    public static final ThingTypeUID THING_TYPE_HELIOS_VENTILATION = new ThingTypeUID(BINDING_ID, "heliosVentilation");
 
-    public static final HashMap<Byte, HeliosVentilationDataPoint> DATAPOINTS;
-
-    private static final Logger LOGGER;
-    static {
-        /* logger is used by readChannelProperties() so we need to initialize logger first. */
-        LOGGER = LoggerFactory.getLogger(HeliosVentilationBindingConstants.class);
-        DATAPOINTS = readChannelProperties();
-    }
     // List of all Channel ids
-    // Channel ids are only in datapoints.properties and thing-types.xml
-
-    /**
-     * parse datapoints from properties
-     *
-     */
-    private static HashMap<Byte, HeliosVentilationDataPoint> readChannelProperties() {
-        HashMap<Byte, HeliosVentilationDataPoint> result = new HashMap<Byte, HeliosVentilationDataPoint>();
-
-        URL resource = Thread.currentThread().getContextClassLoader().getResource(DATAPOINT_FILE);
-        Properties properties = new Properties();
-        try {
-            properties.load(resource.openStream());
-
-            Enumeration<Object> keys = properties.keys();
-            while (keys.hasMoreElements()) {
-                String channel = (String) keys.nextElement();
-                HeliosVentilationDataPoint dp;
-                try {
-                    dp = new HeliosVentilationDataPoint(channel, properties.getProperty(channel));
-                    if (result.containsKey(dp.address())) {
-                        result.get(dp.address()).append(dp);
-                    } else {
-                        result.put(dp.address(), dp);
-                    }
-                } catch (HeliosPropertiesFormatException e) {
-                    LOGGER.error("could not read resource file {}, binding will probably fail: {}", DATAPOINT_FILE,
-                            e.getMessage());
-                }
-            }
-        } catch (IOException e) {
-            LOGGER.error("could not read resource file {}, binding will probably fail: {}", DATAPOINT_FILE,
-                    e.getMessage());
-        }
-
-        return result;
-    }
-
+    public static final String CHANNEL_OUTSIDE_TEMP = "outsideTemp";
+    public static final String CHANNEL_OUTGOING_TEMP = "outgoingTemp";
+    public static final String CHANNEL_EXTRACT_TEMP = "extractTemp";
+    public static final String CHANNEL_SUPPLY_TEMP = "supplyTemp";
+    public static final String CHANNEL_BYPASS_TEMP = "bypass_temp";
+    public static final String CHANNEL_RH_LIMIT = "rh_limit";
+    public static final String CHANNEL_HYSTERESIS = "hysteresis";
+    public static final String CHANNEL_FANSPEED = "fanspeed";
+    public static final String CHANNEL_MIN_FANSPEED = "min_fanspeed";
+    public static final String CHANNEL_MAX_FANSPEED = "max_fanspeed";
 }
