+++ bundles/org.openhab.binding.heliosventilation/src/main/java/org/openhab/binding/heliosventilation/internal/HeliosVentilationHandlerFactory.java
@@ -25,9 +25,10 @@ import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
 import org.eclipse.smarthome.io.transport.serial.SerialPortManager;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * The {@link HeliosVentilationHandlerFactory} is responsible for creating things and thing
@@ -43,10 +44,15 @@ public class HeliosVentilationHandlerFactory extends BaseThingHandlerFactory {
             .singleton(THING_TYPE_HELIOS_VENTILATION);
 
     private @NonNullByDefault({}) SerialPortManager serialPortManager;
+    private final Logger logger = LoggerFactory.getLogger(HeliosVentilationHandlerFactory.class);
 
-    @Activate
-    public HeliosVentilationHandlerFactory(@Reference SerialPortManager serialPortManager) {
+    @Reference
+    protected void setSerialPortManager(final SerialPortManager serialPortManager) {
         this.serialPortManager = serialPortManager;
+    }
+
+    protected void unsetSerialPortManager(final SerialPortManager serialPortManager) {
+        this.serialPortManager = null;
     }
 
     @Override
