--- bundles/org.openhab.binding.heliosventiremote/pom.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.openhab.addons.bundles</groupId>
-    <artifactId>org.openhab.addons.reactor.bundles</artifactId>
-    <version>2.5.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>org.openhab.binding.heliosventiremote</artifactId>
-
-  <name>openHAB Add-ons :: Bundles :: HeliosVentiRemote Binding</name>
-
-</project>
