--- bundles/org.openhab.binding.heliosventiremote/src/main/feature/feature.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<features name="org.openhab.binding.heliosventiremote-${project.version}" xmlns="http://karaf.apache.org/xmlns/features/v1.4.0">
-    <repository>mvn:org.openhab.core.features.karaf/org.openhab.core.features.karaf.openhab-core/${project.version}/xml/features</repository>
-
-    <feature name="openhab-binding-heliosventiremote" description="HeliosVentiRemote Binding" version="${project.version}">
-        <feature>openhab-runtime-base</feature>
-        <feature>openhab-transport-serial</feature>
-        <bundle start-level="80">mvn:org.openhab.addons.bundles/org.openhab.binding.heliosventiremote/${project.version}</bundle>
-    </feature>
-</features>
