--- bundles/org.openhab.binding.heliosventiremote/src/main/resources/ESH-INF/binding/binding.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<binding:binding id="heliosventiremote" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:binding="https://openhab.org/schemas/binding/v1.0.0"
-	xsi:schemaLocation="https://openhab.org/schemas/binding/v1.0.0 https://openhab.org/schemas/binding-1.0.0.xsd">
-
-	<name>HeliosVentiRemote Binding</name>
-	<description>This is the binding for Helios Ventilation Systems KWL EC 200/300/500 Pro. It requires a connection to the RS485 bus used by the original remote controls KWL-FB (9417).</description>
-	<author>Raphael Mack</author>
-
-</binding:binding>
