--- bundles/org.openhab.binding.heliosventiremote/src/main/resources/ESH-INF/i18n/heliosventiremote_xx_XX.properties
@@ -1,17 +1,0 @@
-# FIXME: please substitute the xx_XX with a proper locale, ie. de_DE
-# FIXME: please do not add the file to the repo if you add or change no content
-# binding
-binding.heliosventiremote.name = <Your localized Binding name>
-binding.heliosventiremote.description = <Your localized Binding description>
-
-# thing types
-thing-type.heliosventiremote.sample.label = <Your localized Thing label>
-thing-type.heliosventiremote.sample.description = <Your localized Thing description>
-
-# thing type config description
-thing-type.config.heliosventiremote.sample.config1.label = <Your localized config parameter label>
-thing-type.config.heliosventiremote.sample.config1.description = <Your localized config parameter description>
-
-# channel types
-channel-type.heliosventiremote.sample-channel.label = <Your localized Channel label>
-channel-type.heliosventiremote.sample-channel.description = <Your localized Channel description>
