--- bundles/org.openhab.binding.heliosventiremote/src/main/resources/ESH-INF/thing/thing-types.xml
@@ -1,112 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<thing:thing-descriptions bindingId="heliosventiremote"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:thing="https://openhab.org/schemas/thing-description/v1.0.0"
-	xsi:schemaLocation="https://openhab.org/schemas/thing-description/v1.0.0 https://openhab.org/schemas/thing-description-1.0.0.xsd">
-
-	<!-- Ventilation Thing Type -->
-	<thing-type id="heliosVentilation">
-		<label>HeliosVentilation (KWL)</label>
-		<description>A domestic ventilation system (KWL) from Helios.</description>
-
-		<channels>
-            <channel id="outsideTemp" typeId="outside_temperature" />
-            <channel id="outgoingTemp" typeId="outgoing_temperature" />
-            <channel id="extractTemp" typeId="extract_temperature" />
-            <channel id="supplyTemp" typeId="supply_temperature" />
-            <channel id="bypass_temp" typeId="bypass_temperature" />
-            <channel id="min_fanspeed" typeId="min_fanspeed" />
-            <channel id="max_fanspeed" typeId="max_fanspeed" />
-            <channel id="fanspeed" typeId="fanspeed" />
-            <channel id="rh_limit" typeId="rh_limit" />
-		</channels>
-
-        <config-description>
-            <parameter name="serialPort" type="text" required="true">
-                <context>serial-port</context>
-                <label>RS485 Interface Serial Port</label>
-                <description>The serial port name for the RS485 interfaces. Valid values
-                    are e.g. COM1 for Windows and /dev/ttyS0 or
-                    /dev/ttyUSB0 for Linux.</description>
-            </parameter>
-
-            <parameter name="pollPeriod" type="integer" required="true" min="0">
-                <label>Poll Period (sec.)</label>
-                <description>The Poll Period in Seconds use 0 for no polling.</description>
-                <default>60</default>
-            </parameter>
-        </config-description>
-	</thing-type>
-
-    <channel-type id="outside_temperature" advanced="false">
-        <item-type>Number</item-type>
-        <label>Outside Temperature</label>
-        <description>Temperature measured in the outdoor air flow.</description>
-        <category>Temperature</category>
-        <state pattern="%d °C" readOnly="true" />
-    </channel-type>
-    
-    <channel-type id="extract_temperature" advanced="false">
-        <item-type>Number</item-type>
-        <label>Extract Temperature</label>
-        <description>Temperature measured in the extract (indoor, room temperature) air flow.</description>
-        <category>Temperature</category>
-        <state pattern="%d °C" readOnly="true" />
-    </channel-type>
-
-    <channel-type id="supply_temperature" advanced="false">
-        <item-type>Number</item-type>
-        <label>Supply Temperature</label>
-        <description>Temperature measured in the supply (incoming) air flow.</description>
-        <category>Temperature</category>
-        <state pattern="%d °C" readOnly="true" />
-    </channel-type>
-
-    <channel-type id="outgoing_temperature" advanced="false">
-        <item-type>Number</item-type>
-        <label>Outgoing Temperature</label>
-        <description>Temperature measured in the outgoing air flow.</description>
-        <category>Temperature</category>
-        <state pattern="%d °C" readOnly="true" />
-    </channel-type>
-
-    <channel-type id="fanspeed" advanced="false">
-        <item-type>Number</item-type>
-        <label>Fanspeed</label>
-        <description>Fanspeed.</description>
-        <category>HVAC</category>
-        <state pattern="%d" readOnly="false" />
-    </channel-type>
-
-    <channel-type id="min_fanspeed" advanced="true">
-        <item-type>Number</item-type>
-        <label>Min Fanspeed</label>
-        <description>Minimal Fanspeed.</description>
-        <category>HVAC</category>
-        <state pattern="%d" readOnly="false" />
-    </channel-type>
-
-    <channel-type id="max_fanspeed" advanced="true">
-        <item-type>Number</item-type>
-        <label>Max Fanspeed</label>
-        <description>Maximum Fanspeed.</description>
-        <category>HVAC</category>
-        <state pattern="%d" readOnly="false" />
-    </channel-type>
-
-    <channel-type id="bypass_temperature" advanced="true">
-        <item-type>Number</item-type>
-        <label>Cell Bypass Temperature</label>
-        <description>Bypass temperature to disable the bypass function if outside temperature is below this threshold even if ventilation system is in summer mode.</description>
-        <category>Temperature</category>
-        <state pattern="%d °C" readOnly="false" />
-    </channel-type>
-
-    <channel-type id="rh_limit" advanced="true">
-        <item-type>Number</item-type>
-        <label>RH limit</label>
-        <description>Limit for relative humidity sensor.</description>
-        <category>Humidity</category>
-        <state pattern="%f %rh" readOnly="false" />
-    </channel-type>
-</thing:thing-descriptions>
